backend/requirements.txt
Flask
Flask-Cors
pydantic
python-slugify
ruamel.yaml
pytest
